pyproject.toml
[project]
name = "health-insurance-agent-2"
version = "0.1.0"
description = "HealthGuard Agentic RAG - An intelligent multi-modal assistant for US Health Insurance members"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "streamlit>=1.38.0",
    "groq>=0.11.0",
    "openai>=1.30.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",
    "pandas>=2.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "tabulate>=0.10.0",
]

[project.scripts]
setup-data = "backend.scripts.setup_data:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend"]
